more work on python decorator and bindings/linkage
